Refactor HumanTimeDiffWithDateTip to be a component instead of a funciton of helpers.ts
