+++ frontend/pages/hosts/details/cards/About/About.tsx
@@ -1,13 +1,10 @@
 import React from "react";
 
 import ReactTooltip from "react-tooltip";
+import HumanTimeDiffWithDateTip from "components/HumanTimeDiffWithDateTip";
 
 import { IHostMdmData, IMunkiData, IDeviceUser } from "interfaces/host";
-import {
-  humanHostLastRestart,
-  humanHostEnrolled,
-  humanHostTimeDiffWithDateTip,
-} from "utilities/helpers";
+import { humanHostLastRestart } from "utilities/helpers";
 
 interface IAboutProps {
   aboutData: { [key: string]: any };
@@ -22,7 +19,6 @@ const About = ({
   deviceMapping,
   munki,
   mdm,
-  wrapFleetHelper,
 }: IAboutProps): JSX.Element => {
   const renderSerialAndIPs = () => {
     return (
@@ -155,21 +151,20 @@ const About = ({
         <div className="info-grid__block">
           <span className="info-grid__header">Added to Fleet</span>
           <span className="info-grid__data">
-            {/* {wrapFleetHelper(humanHostEnrolled, aboutData.last_enrolled_at)} */}
-            {humanHostTimeDiffWithDateTip(
-              aboutData.last_enrolled_at ?? "Unavailable"
-            )}
+            <HumanTimeDiffWithDateTip
+              timeString={aboutData.last_enrolled_at ?? "Unavailable"}
+            />
           </span>
         </div>
         <div className="info-grid__block">
           <span className="info-grid__header">Last restarted</span>
           <span className="info-grid__data">
-            {humanHostTimeDiffWithDateTip(
-              humanHostLastRestart(
+            <HumanTimeDiffWithDateTip
+              timeString={humanHostLastRestart(
                 aboutData.detail_updated_at,
                 aboutData.uptime
-              )
-            )}
+              )}
+            />
           </span>
         </div>
         <div className="info-grid__block">
