+++ frontend/pages/hosts/details/cards/HostSummary/HostSummary.tsx
@@ -5,11 +5,8 @@ import TooltipWrapper from "components/TooltipWrapper";
 
 import Button from "components/buttons/Button";
 import DiskSpaceGraph from "components/DiskSpaceGraph";
-import {
-  humanHostMemory,
-  wrapFleetHelper,
-  humanHostTimeDiffWithDateTip,
-} from "utilities/helpers";
+import HumanTimeDiffWithDateTip from "components/HumanTimeDiffWithDateTip";
+import { humanHostMemory, wrapFleetHelper } from "utilities/helpers";
 import getHostStatusTooltipText from "pages/hosts/helpers";
 import StatusIndicator from "components/StatusIndicator";
 import IssueIcon from "../../../../../../assets/images/icon-issue-fleet-black-50-16x16@2x.png";
@@ -212,9 +209,13 @@ const HostSummary = ({
               {deviceUser ? "My device" : titleData.display_name || "---"}
             </h1>
             <p className="last-fetched">
-              {titleData.detail_updated_at
-                ? humanHostTimeDiffWithDateTip(titleData.detail_updated_at)
-                : "Unavailable"}
+              {titleData.detail_updated_at ? (
+                <HumanTimeDiffWithDateTip
+                  timeString={titleData.detail_updated_at}
+                />
+              ) : (
+                "Unavailable"
+              )}
             </p>
             {renderRefetch()}
           </div>
